:art: 重构课表框架结构
- 将 AbstractHomeCourseFrame 重命名为 AbstractCourseFrame 并移动到 frame 包
- 将 AdaptiveHomeCourseFrame 重命名为 AdaptiveCourseFrame 并更新相关引用
- 移除 IAdaptiveHomeCourseFrame 接口，直接使用实现类
- 将各类装饰器视图模型从 home.item.decoration 移动到 frame.decoration
- 将各类项目工厂从 home.item 移动到 frame.item
- 将底部工作表对话框相关文件从 home.dialog 移动到 dialog.item
- 将头部组件从 home.compose 移动到 frame.header
- 删除不再使用的 HomeLinkLessonDataProvider 和 HomeSelfLessonDataProvider
- 将 MobileHomeCourseFrame 移动到 frame 包并更新依赖
- 将底部工作表相关文件从 home.compose 移动到 frame.bottomsheet
