--- cyxbs-pages/course/api/src/commonMain/kotlin/com/cyxbs/pages/course/api/IAdaptiveHomeCourseFrame.kt
@@ -1,18 +1,0 @@
-package com.cyxbs.pages.course.api
-
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-
-/**
- * .
- *
- * @author 985892345
- * @date 2025/9/22
- */
-interface IAdaptiveHomeCourseFrame {
-
-  @Composable
-  fun HomeCourseContent(
-    modifier: Modifier,
-  )
-}
